Fix Random::rand{32,64} when RNG has less than 32 bits of entropy
Summary:
`rand32()` and `rand64()` assume that the passed RNG always has at least 32 bits of entropy, which is not always the case (for example `std::minstd_rand` only has 31), so fall back on `uniform_int_distribution` in that case, and optimize `rand64()` when the passed generator has 64 bits of entropy.

NOTE: This is a resubmit of D37851562 with a workaround for a MSVC bug, replacing `RNG::max() + 1` with `~RNG::max()` (see https://www.internalfb.com/diff/D37851562?transaction_fbid=401587391951037). `build-ovr-server` is now green.

Reviewed By: Gownta

Differential Revision: D37927090

fbshipit-source-id: 69aacd0afc64f06d5f2839799ae44ddc6db43b89