Remove intrusive linked list overload of consumeChannelWithCallback
Summary:
The intrusive linked list overload of consumeChannelWithCallback is tricky to use, since ChannelCallbackHandleList is not thread-safe. This means that all operations affecting the list must run on the same SerialExecutor.

There are no remaining uses of this overload, so this diff removes it. Similar behavior can be achieved with ChannelProcessor.

Reviewed By: aary

Differential Revision: D37972588

fbshipit-source-id: ee5b443ba4ccbcb4365c6fb1059662b377ec9d0d