Optimize Random::randDouble01
Summary:
The compiler is not able to optimize away a lot of the generality of `std::generate_canonical()` from libstdc++, in particular a call to `std::log()` on the generator range is not constant-folded, and results in very slow calls.

It's very quick to derive a random double in `[0, 1)` from a uniform 64-bit integer just by scaling, so we can do that, which turns out to be about 15x faster in microbechmarks.

Reviewed By: Orvid

Differential Revision: D37851561

fbshipit-source-id: ce439e939c88e03fa52ac5e425f14e4528373afc