ThrottledLifoSem
Summary:
In online services, thread pools are generally operated in a regime where there are frequently idle threads, to guarantee small queueing delay. This implies that most tasks submitted immediately wake up a thread to be executed, incurring a wake-up cost of a few microseconds of CPU time (split between the waker and the waiter).

If such tasks are very cheap (order of hundreds of nanoseconds) the wake-up overhead becomes very significant. This is not uncommon in coroutine and Future heavy code, where many tasks do just bookkeeping of some shared state, or only forward work to a different executor.

This diff introduces `ThrottledLifoSem`, a drop-in replacement for `LifoSem` which only wakes up a thread if there have been outstanding posts for a configurable amount of time. If a burst of small tasks is submitted, it is more likely to be executed by threads that are already running, instead of waking up all the idle threads.

When plugged in a thread pool, this roughly simulates a continuous resizing where the pool size is incremented only when the queueing delay crosses the `wakeUpInterval` threshold, but the resizing is done without actually shutting down or creating threads.

Reviewed By: Gownta

Differential Revision: D37267339

fbshipit-source-id: f97595b45f90ddcf3a41d6d77e05ef39a594f20c