AsyncUDPSocket allow nullptr for evb on construction
Summary:
There's no technical reason for an `EventBase` to be required on construction given that it's already possible to have an `AsyncUDPSocket` with no `EventBase` attached.

Change is necessary to enable unit test where the socket has no `EventBase` on construction; test is blocked by non-conditional call to `eventBase_->dcheckIsInEventBaseThread`.

Differential Revision: D37745083

fbshipit-source-id: 81071e8beadf9edf2d8f3d0b788ff4e861d75b12