Make folly RCU lock() and unlock() noexcept
Summary:
The folly RCU implementation currently can throw on the lock() path due to
performing a lazy allocations for the TLS epoch | reader count
std::atomic<uint64_t> for a given domain. Folly often treats allocation
failures as terminal, and the folly RCU TS2 spec requires lock() and
unlock() to be noexcept, so this diff marks them as noexcept. Note
that we could still possibly throw on these paths, but we intentionally
want allocation failures here to be terminal to match the spec, so we
mark them as noexcept regardless.

Reviewed By: davidtgoldblatt

Differential Revision: D37254592

fbshipit-source-id: ce5cd13a678b9629dbc2c243f7472a97a82c75e4