Synchronized LockedPtr Make parent() public
Summary: Make LockedPtr's parent() function public. We have a function that detaches a thread and that detached thread needs to eventually aquire the lock. By making this public, the SynchronizedType*  can be passed to the thread so that it can get the lock itself. If we simply passed the lockedState to the thread it causes tsan to flag it as the lock is unlocked from a different thread.

Reviewed By: Orvid

Differential Revision: D38639576

fbshipit-source-id: 9cab57ee5442c7f321a037c9f9774c081c7bb8e3