cut exception_wrapper on-heap storage variant
Summary:
Cut `exception_wrapper` onheap storage variant as a step in revising `exception_wrapper`.

This takes advantage of `std::exception_ptr` being sufficient for storing unthrown exceptions as well as thrown exceptions as of libstdc++7.

The on-heap storage variant previously existed as an optimization around the thrown variant, which is backed by `std::exception_ptr`. The primary motivations were the lack of an efficient way to create a `std::exception_ptr` in prior implementations of the standard library. While the standardd library also lacks efficient ways to inspect the stored object of a `std::exception_ptr`, the thrown variant has historically solved that case by caching an extra pointer to accelerate that inspection and now solves that case by reaching into the abi/runtime.

Reviewed By: Gownta

Differential Revision: D26377019

fbshipit-source-id: f89651d32d27ebb8c039d3e74e5a25df3fbf2733