Inline a method to save two hot load instructions.
Summary:
Inline coalesceWithHeadroomTailroom into coalesce to save two loads. The
profile information shows that we usually don't enter the if inside the
'tailroom' function, and the compiler can't sinkt the two method calls, so the
two loads are wasted. Sinking the two loads/method-calls into the the body of
the IF should allow us to avoid the cost of the hot loads.

Differential Revision: D39221756

fbshipit-source-id: 6ed05585f2247d177642f147e9cab4d5445d716f