Inline a method to save two hot load instructions. (#1855)
Summary:
Pull Request resolved: https://github.com/facebook/folly/pull/1855

Inline coalesceWithHeadroomTailroom into coalesce to save two loads. The
profile information shows that we usually don't enter the if inside the
'tailroom' function, and the compiler can't sink the two method calls, so the
two loads are wasted. Sinking the two loads/method-calls into the the body of
the IF should allow us to avoid the cost of the hot loads.

Reviewed By: ot, philippv

Differential Revision: D39221756

fbshipit-source-id: 98e4f222c04f3fbf98c1de32bb7cdb0f2d9719f1