Break deadline ties by submission order in EDFThreadPoolExecutor
Summary:
Deadline ties are currently scheduled in unspecified order. This can be counterintuitive, given that other folly executors with priorities schedule FIFO within the same priority, and `EDFThreadPoolExecutor` is often used through the `deadlineExecutor()` adapter, which does not behave as a FIFO executor.

We already serialize enqueues for a given deadline, so adding a tie-breaker based on enqueueing order requires no additional synchronization cost.

Reviewed By: philippv

Differential Revision: D39341998

fbshipit-source-id: 9d491cd1d451182f35ffb472d5fb19fe72b2e685