Make the fast path of ThrottledLifoSem::post() lock-free
Summary:
Most calls to `post()` should not cause a wakeup (this is the main goal that `ThrottledLifoSem` tries to accomplish) but the current implementation still acquires a lock to determine whether a wakeup is needed.

By fitting all the relevant state in a single word, we can avoid acquiring the lock unless it's very likely that a wake-up will occur. In particular this avoids locking entirely when there are no waiters (for example if the thread pool that uses the semaphore is saturated).

With this change, `ThrottledLifoSem` with `wakeUpInterval = 0` should be amenable to be a safe replacement to `LifoSem` for virtually all workloads.

For this reason, also change the default `wakeUpInterval` to `0`. All existing uses of `ThrottledLifoSem` provide an explicit value anyway.

Reviewed By: Gownta

Differential Revision: D39419142

fbshipit-source-id: 6a17a4db17a774bcd857608736ae21f8fd700628