Revert back to a spinlock in DigestBuilder
Summary:
Counters are used in several latency-critical codepaths, for which a preemption (either through `yield` or a futex) can be very damaging.

D31654923 (https://github.com/facebook/folly/commit/580d5f19e760db938a92b9d2b1d969d5955ab0b3) switched the spinlock to a `SharedMutex`, which is very eager to yield the CPU on contention, but that comes at the cost of tail latency. Instead, revert to a spinlock, but try to remove sources of contention by making the critical section very unlikely to be preempted.

Note that the mutex is per-cpu, so contention can happen only if we have a stale `cachedCurrent()`, or the thread migrates while in the critical section, or `build()` is running (but the critical section of that is also very small).

Reviewed By: terrelln

Differential Revision: D39557800

fbshipit-source-id: 7bfb05bb9456830c670d7127e2b022851586394f