Avoid calling clock::now in the common case.
Summary:
Avoid calling clock::now in the common case.

Clock::now() can be an expensive call (a call to libc's __clock_gettime). PGO
data shows that the first check almost always succeeds, which makes the call to
Clock::now redundant. This patch peels one iteration that calls f() before
calling clock::now(). Notice that a very common use of this API is a function
that performs a single load in the function, so timing is not a major concern.

Differential Revision: D39611860

fbshipit-source-id: 105ce07995bdc96627aa8b3011dc565645a498c6