add options to kill the child in the Subprocess destructor
Summary:
Add an option to folly::Subprocess to automatically kill and wait on the child
process if it is still running when the folly::Subprocess destructor is
invoked.

Without this setting, callers have to take extra care to ensure that they
correctly terminate and wait for the process on all code paths, since this
would not happen automatically for them with RAII.  Getting this logic
correct is non trivial.  Some call sites were instead using the
`allowDestructionWhileProcessRunning()` option, which is undesirable, since
this option actively leaks resources, resulting in zombie processes.

It is still preferable for callers to terminate and wait for the process on
their own in most cases, rather than letting this happen in the destructor:
terminating the process in the destructor may block for some amount of time,
and it also provides no option to check for any errors.  Nonetheless,
providing this behavior as an option still allows using RAII to rely on the
process to be correctly cleaned up on all code paths.

Reviewed By: luciang

Differential Revision: D39790289

fbshipit-source-id: d34d2071036ace19c1266a91f11e4d492233d649