accept EWOULDBLOCK & log other errnos
Summary:
On Windows we see these CHECKs firing unexpectedly.

The strong suspicion is that `_read` is returning -1 and setting errno to EWOULDBLOCK. I verified that EWOULDBLOCK and EAGAIN are different numerical values on Windows whereas I believe they are typically the same value on Linux platforms.

In case that suspicion is incorrect, this diff modifies the CHECKs to log the values of `result` and `errno` when a failure occurs.

Reviewed By: yfeldblum

Differential Revision: D29348730

fbshipit-source-id: 617bbac8a1235fbdc923aa0b64a4670f1fde13d4