Fix double Baton wait in ThrottledLifoSem
Summary: I misread the `Baton` documentation and, contrary to `try_wait()`, `try_wait_until()` is destructive even when it returns `false`, so we cannot call `wait()` after a timeout. The fix is to switch to `SaturatingSemaphore`. In fact, `LifoSem` uses that too for the same reason, but I overlooked that.

Reviewed By: Gownta

Differential Revision: D40306917

fbshipit-source-id: da7786fa31de81f5f8053307f5d3e37939410d5e