Fix cubic test runtime in MemsetTest
Summary:
memset_test has three for-loops: alignment, length, and then memset. For unalignedBuffer, this cubic behavior is taking painfully long to run - so long that test frameworks are, rightly, timing it out. Reduce the length parameter for unaligned testing.

Aside: I compared runtimes for folly's memset with regular memset. Regular memset was faster for unaligned, but not too much so.

Reviewed By: yfeldblum

Differential Revision: D40243712

fbshipit-source-id: 2362712d55c45769b7d758c9f314ebc662d1f88a