Use nallocx instead of naiveGoodMallocSize
Summary:
This reverts the behavior of folly::goodMallocSize to before D38265684 (https://github.com/facebook/folly/commit/2b4c73f78cfe2745a76b1d6625d3fb38f1ce667e).
This is done because of S303225 which suggests that something changed related
to the introduction of naiveGoodMallocSize. The function naiveGoodMallocSize
remains available, it just won't be used by goodMallocSize when we are using
jemalloc.

Reviewed By: interwq

Differential Revision: D40596988

fbshipit-source-id: 512c5338e6df804585c1728d151f8ab5c90cce7a