Optimize folly::to<> for 128-bit integers to string
Summary:
The current version does one 128-bit division for each digit outside of the 64-bit range, which is extremely slow.
Instead we can rewrite the integer in base-10^19 (requiring at most 2 128-bit divisions), so that each component fits in a 64-bit integer and we can use 64-bit arithmetic for most of the divisions.

Reviewed By: Gownta

Differential Revision: D40715247

fbshipit-source-id: 17116e8a3fd2940f4eaa4106229265a573a54799