Always set a thread name in FunctionScheduler
Summary:
The name for the scheduling thread is specified with `FunctionScheduler::setThreadName`, which is not always called (or may even be called too late!), in which case the thread name is inherited from thread that calls `start()`, leading to confusing attribution.

Use a default name instead, as we do in several other executor-like primitives.

Reviewed By: philippv, luciang

Differential Revision: D40797728

fbshipit-source-id: 2e3d2516b0dc2fd07b6c2783de30797c8c2d2f43