build speed tweak for CacheLocality: no constinit table of relaxed_atomic
Summary:
The compiler must evaluate the constructor for each element of the table, which is observed to cost compile time. Switch to a constinit array of raw bytes instead and always use `atomic_ref`s to access the table in order to avoid valuating all of the `relaxed_atomic` constructors at runtime.

```name=master
$ foundation/scripts/run-perf-compile --quiet --syntax --platform platform010 --which perf folly/concurrency/test/CacheLocalityTest.cpp
     9,275,503,791      instructions:uP
```
```name=branch
$ foundation/scripts/run-perf-compile --quiet --syntax --platform platform010 --which perf folly/concurrency/test/CacheLocalityTest.cpp
     4,577,023,094      instructions:uP
```

Reviewed By: Gownta

Differential Revision: D40504017

fbshipit-source-id: b4cb22b27ed8f31787b0f860736be0e4b56e38da