Give AtomicCoreCachedSharedPtr slightly stronger reclamation guarantees
Summary:
`AtomicCoreCachedSharedPtr` uses hazard pointers to manage the `shared_ptr` cache, so on `reset()` the old references are released in background threads. Releasing them synchronously (waiting for any outstanding readers) would be expensive due to the design of hazard pointers, but we can at least ensure that they are all released when `AtomicCoreCachedSharedPtr` is destroyed.

See for example in the changed test how this makes the semantics a bit more deterministic.

Reviewed By: Gownta

Differential Revision: D40721918

fbshipit-source-id: 5c31ae885687775e306ae98dabc8994178058bb6