Optimize SerialExecutor when the queue is non-empty
Summary: Port to `SerialExecutor` an optimization already present in `StrandExecutor`: instead of scheduling multiple workers and making one of them win the race to process the queue, ensure that only one worker is scheduled at any given time. This saves the scheduling of useless tasks on the parent executor when the queue is non-empty.

Reviewed By: luciang

Differential Revision: D42069420

fbshipit-source-id: cd68a29a9212766307a676d5e80ed48e88dc0d70