split useful structures out of IoUringBackend.h
Summary:
Including IoUringBackend.h forces including of liburing.h.
This tends to conflict with lots of existing code that happens to use the same symbols as the kernel headers have defined - eg BLOCK_SIZE

Instead provide a minimal include file that can be used instead, and does not pull in all the kernel headers.

Reviewed By: dmm-fb

Differential Revision: D40550358

fbshipit-source-id: 48b8a6b72640012505a800e15501677ae047e00d