Shrink F14 maps
Summary: It would be nice if F14 maps were smaller. This diff steals 1 byte of the 8-byte size word to hold the base-2 logarithm of the chunk count and eliminates the 8-byte storage for `chunkMask_`. Thanks to the BZHI instruction (which computes `x & ((1UL << shift) - 1)` given `x` and `shift`) , the find path is roughly the same as before.

Reviewed By: yfeldblum

Differential Revision: D40367479

fbshipit-source-id: 8d9811ee748245a2d4a9c31e6acace8a88ff7e01