CoreAllocator improvements
Summary:
Make `CoreAllocator` not templated on `kMaxSlots`, by mapping back to "locality indexes" so we can share the arenas among different use cases.

In practice, there is only one `kMaxSlots`, as this is only used by `CoreCachedSharedPtr`, but the eventual goal is to move all the implementation details to the cpp file. This happens in the next diff.

Also, when running under ASAN make deallocation scrub the buffer. The reason is that we leak the arenas at shutdown, and there might be pointers left in the deallocated objects; those would prevent LSAN from detecting a leak.

Reviewed By: luciang

Differential Revision: D41531575

fbshipit-source-id: 4b166f41f8a9d25689d9f2cd5d75a849ca17947b