Add SerializedExecutor interface
Summary:
We don't have an executor interface that guarantees non-concurrent task execution, `SequencedExecutor` only guarantees that if the corresponding `add()`s are sequenced too.

OTOH we have multiple implementations of executors that provide that guarantee, 2 in folly (`SerialExecutor` and `StrandExecutor`) and others internally, and it would be useful to be able to type-erase them so the implementation can be selected at runtime.

Differential Revision: D42690040

fbshipit-source-id: 2b2b6b6f39090a44f0476a1ad338d4fc15f07316