Add support for constexpr, function pointers, and member pointers to `folly::overload`
Summary:
The implementation of `folly::overload` required that the invocables passed in were class types. This extends the implementation to support function pointers, member function pointers, and member data pointers.

This also adds `constexpr` support.

Finally, `overload()` (attempting to create an empty overload set) previously resulted in a compile-time error. Now it extends it in the obvious way -- it creates an empty overload set.

Improve documentation to explicitly list the set of types supported by `overload`.

Reviewed By: ispeters

Differential Revision: D42510591

fbshipit-source-id: bd47b3547078f43dfff2d0cc08273c01247948ea