logging: convince XCHECK(false) to be noreturn
Summary:
`XCHECK(false)` was not being marked noreturn because the `XLOG_IS_ON_IMPL`
check was preventing the compiler from figuring out that the entire thing was
noreturn.

For `FATAL` log messages only, we make `XLOG_IS_ON_IMPL` statically evaluate to
`true`. This is not a behavior change afaict because we do not allow people to
disable `XLOG(FATAL)` log messages (although I would like confirmation here,
otherwise this would be a pretty crazy change!).

Reviewed By: simpkins

Differential Revision: D42907276

fbshipit-source-id: edb6e4b166823b1c1aab72127825205e0752f44b