fix build errors in Base64PlatformTest.cpp on Windows
Summary:
This code was trying to assign a `std::array::iterator` to a `uint8_t*`.
In libstdc++ the `std::array::iterator` type happens to be a simple pointer,
but this is not true on Windows, resulting in a build failure.

This updates the code to use `std::array::data()` to get a pointer to the
array data, and not `std::array::begin()`.

Reviewed By: DenisYaroshevskiy

Differential Revision: D42948330

fbshipit-source-id: cbd90d80aed9443af8a349111c06619fb0a2c901