Expose name prefix in ThreadFactory
Summary:
`ThreadPoolExecutor` exposes the thread name prefix through `getName()`, but the thread names are determined by the thread factory, so the implementation assumes that the factory is (a derived class of) `NamedThreadFactory`.

This breaks when using custom thread factories or compositions built with `InitThreadFactory`. Instead, this diff exposes `getNamePrefix()` in `ThreadFactory`. This does not place significant additional burden in implementing thread factories, as `InitThreadFactory` is usually enough to implement the desired semantics.

Reviewed By: philippv, luciang

Differential Revision: D43155851

fbshipit-source-id: e9eeaa6ef5b4be91a082d66da14565ef30aff4b9