folly::detail::simdSplitByChar (#1947)
Summary:
Pull Request resolved: https://github.com/facebook/folly/pull/1947

X-link: https://github.com/facebook/hhvm/pull/9345

Simd optimizied version of splitting strings by char for x86 SSE2/AVX2 and AARCH64.

It's not enabled yet - I plan to let fuzzing run for a few days first.

As far as the decision to make the implementation out of line - I think it's for the best, you don't want to inline that huge piece of code. And the compile time should be a bit faster.. But if you think that pre-generating those copies is worse - I can inline instead.

Differential Revision: D43757774

fbshipit-source-id: d777eac38538b934c009f72a837fccf901c0e474