fix folly exception tracer
Summary:
The symbols in folly's ExceptionTracerLib.cpp were designed to wrap the ones by the same name in libstdc++. This was fine when libstdc++ was being dynamically linked with folly, but causes duplicate symbol errors when libstdc++ is statically linked. To get around this we can use the `--wrap` linker flag which works as follows:
```
Use a wrapper function for symbol.
Any undefined reference to symbol will be resolved to __wrap_symbol.
Any undefined reference to __real_symbol will be resolved to symbol.
This can be used to provide a wrapper for a system function.
The wrapper function should be called __wrap_symbol.
If it wishes to call the system function, it should call __real_symbol.
```

Reviewed By: igorsugak

Differential Revision: D44158339

fbshipit-source-id: 56c3dde6546c816e2c988f07aaade9b0dd2c364a