Zero out `msghdr` to avoid ancillary data bugs with pre-received data
Summary:
When `AsyncFdSocket` (which is the only consumer of the `ReadAncillaryCallback` API for now) is used in Thrift, it can end up going through "peeking manager", which is a gadget that first reads 13 bytes one socket, uses that to determine the connection type, then moves the socket FD to a new class, and passes the previously read data to the socket via `setPreReceivedData`.

We can thus end up on the pre-received data code path in `performReadMsg` with `readAncillaryCallback_` set ... and yet that code path would leave `msg_control`, `msg_controllen`, and `msg_flags` uninitialized (i.e. garbage). That would in turn lead to failures in `AsyncFdSocket`'s `ancillaryData()` callback.

It seems safest to always zero-initialize `msghdr`. This avoids having the expectation that every possible code path through `performReadMsg` (of which there are three in the base `AsyncSocket` and more in child classes) correctly sets ancillary data. Defaulting to zeros signals "there's no ancillary data", which is a good behavior in all code paths that don't `recvmsg`.

Reviewed By: dmm-fb

Differential Revision: D44568001

fbshipit-source-id: ec9f1e747ed4f238ae36ded641006008bac38962