Add unique pointer type for EVP_ENCODE_CTX (#1975)
Summary:
Pull Request resolved: https://github.com/facebook/folly/pull/1975

Add a unique pointer wrapper for EVP's base64 encoding/decoding function.

The `EVP_ENCODE_CTX` is included in all openssl version >= 1.1.1.
We don't do openssl version check (as in the case of `EVP_PKEY`) for it, because it is included in a same openssl version as ` EVP_MD_CTX` (line 93), which doesn't do this check either.

Ref: https://www.openssl.org/docs/man1.1.1/man3/EVP_ENCODE_CTX_new.html

Reviewed By: Orvid

Differential Revision: D44877508

fbshipit-source-id: 3581f8fd06201931ef8777721d7a9a79cb05cd55