RFC: [folly][Expected] std::expected compatibility: allow non-copy constructible error types
Summary:
`std::expected::value() &&` is documented to throw `std::move(error())`, which allows for non-copy constructible error types.

This `throw` in `requireValue()` is the only reason we require errors to be copy-constructible in `folly::Expected` currently, so this change makes the behaviour compatible with `std::expected`.

Rationale:
Using `Expected` with non-copy constructible errors and values will alert the users to unintentional copies which could otherwise cause performance regressions.

Reviewed By: Orvid

Differential Revision: D42653289

fbshipit-source-id: 0064eb88e6a76ac425350b64e24ab2a23f81dd43