allow Baton::ready after Baton::wait
Summary:
`Baton::ready` was permitted only before a call to `Baton::wait`, just like `Baton::try_wait` was. This restriction makes more sense for `try_wait` but less sense for `ready`. So we lift the restriction.

`SaturatingSemaphore` and `fibers::Baton` naturally have this relaxation already.

Also fix `try_wait` not to be `const`; the waiting member functions are non-`const` and so `ready` is provided as `const`.

Differential Revision: D45451985

fbshipit-source-id: f194c8ed8eaf30f868e74ff03daf305d6faa6127