Add `popNextReceivedFds` using `ReadAncillaryDataCallback`
Summary:
The "receive" part of `AsyncFdSocket`. Batches of FDs are queued up inside the socket, and can be retrieved FIFO.

Design note: Hypothetically, it could be useful to return FDs as part of `ReadCallback::readDataAvailable`, or analog. This would capture the exact correspondence between "data bytes" and "FDs", and avoid the messy-to-implement and tricky to use API exposed by `ReadAncillaryDataCallback`. However, there are hundreds of extant `AsyncTransport::ReadCallback`s, so the implication of updating or extending this API are considerably larger than bolting on this "queueing" API. And, the current queue-based approach is perfectly adequate for use in Thrift Rocket transport.

Reviewed By: AkramaMirza

Differential Revision: D43935125

fbshipit-source-id: 27cb419c41fdc776d42364f1be2eac39616fe813