Add writeStarting() callback
Summary:
This diff adds the `writeStarting` callback to `AsyncWriter` to indicate when AsyncSocket is ready to start writing a write to the socket. It signals that all previous write have either succeeded or failed and the socket is ready to attempt the current write operation. The callback triggers once per write, even if the write is subsequently split, e.g., through prewrite requests.

One possible use for this callback: the application gets notified when the write is starting and can unambiguously get the raw byte offset at the beginning of the write. Without this callback, the application could get the offset only when calling write(). This would not guarantee an accurate offset as bytes from previous writes may not have yet been written.

Reviewed By: bschlinker

Differential Revision: D44240877

fbshipit-source-id: bb7bc54687db5acd7ee01c81a18a0a044687588b