Move warning disables to outside of unsafe_default_initialized_cv member
Summary:
According to MSVC doc [1], "warning numbers in the range 4700-4999 are associated with code generation. For these warnings, the state of the warning in effect when the compiler reaches the function definition remains in effect for the rest of the function." Therefore, the warning disables inside the `unsafe_default_initialized_cv` member function should be outside of the function.

[1] https://learn.microsoft.com/en-us/cpp/preprocessor/warning

Reviewed By: Orvid, 8Keep

Differential Revision: D45609862

fbshipit-source-id: 7cf28fe9c0e6da1395f03d1ede3fe27ba7f5595b