Support inlining in collect()
Summary:
collect() (as opposed to collectAll()) does not inline callbacks
correctly. Fix it to correctly inline with two separate fixes. First, pass
InlineContination::permit to setCallback_() when attaching a callback to the
input futures. Second, capture the keep alive for the last completing future
and use that keep alive when setting the result with setTry() to correctly
propagate the completing executor so the thenImpl() machinery correctly inlines
later

Differential Revision: D45496500

fbshipit-source-id: 84c9305b9b98924dd7714e77dfa052ae52b99172