Implement `folly::is_allocator` trait
Summary:
The high level idea for this diff stack is impelemnt deduction guides support
for `folly::F14*` containers and re-use the same building blocks to
re-implement deduction guides for `multifeed::QuickHash*` containers.

This is a first diff in a series with implement of `folly::is_allocator` trait.

Based on [libstdc++][1] and [libc++][2] implementations seems it enough to
check for:
* `value_type` type alias,
* `allocate` method.

[1]: https://github.com/gcc-mirror/gcc/blob/057e537e2a6f0fead3ebb503c8dcc08c9c6491ec/libstdc%2B%2B-v3/include/bits/alloc_traits.h#L842-L853
[2]: https://github.com/llvm/llvm-project/blob/3ce72cd5f6d584d83d47e54edef5c132a6c3703d/libcxx/include/__type_traits/is_allocator.h#L24C1-L32

Reviewed By: Gownta

Differential Revision: D45865403

fbshipit-source-id: 6139dfb33a6742ece1a8391c7c11fc57992487ad