Split prehashing and prefetching in F14
Summary:
Prehashing and prefetching should be independent: we may want to prehash without prefetching (to store the token or batch operations on the same key), or we may want to prefetch later (on a stored token). There is no benefit in bundling these operations in the same method.

This diff splits `prehash()` into two self-documenting methods, `prehash()` and `prefetch()`.

Reviewed By: terrelln

Differential Revision: D46018950

fbshipit-source-id: f6775a0d63495f8e2bb3322eaaab7c5a2ae31979