Fixing gcc build (#2009)
Summary:
Pull Request resolved: https://github.com/facebook/folly/pull/2009

Apparently this change caused some warning issues on gcc.
There was a shadow warning - that's easy.
The "loosing attributes" warning is a bit trickier, but if you look for it, you will see it disabled everywhere.

It's basically saying that some attributes (like alignas) will be ignored. And __m128 are done with attributes, though these ones are not ignored.

Alternative is to wrap intrinsics but that's too much.

Reviewed By: Orvid

Differential Revision: D46298812

fbshipit-source-id: 6cc64ea846fcbf881f7e2a04cddd120adcaa42b9