Test send unbatching & fragment handling for FDs
Summary:
This covers the code path where the server handler is sending FDs back to the client (D43935124 and  D43935129). The primary goal is to exercise the FD-based unbatching code -- that is, in the presence of FDs, we must break up the already-batched write into several socket writes. Here's what it checks:
- Separate messages carrying FDs are separated into separate socket invocations.
- Messages without FDs continue being batched together, and are batched with the subsequent FD-carrying message (if there is one in the batch).
- The socket is passed FDs together with the last fragment of a message.
- All the data & FDs enqueued into a batch are sent.
- The right data matches to the right FD in the socket code.
- FD passing also works with batching disabled.

Reviewed By: AkramaMirza

Differential Revision: D45507928

fbshipit-source-id: 729005efc041d10b594e0007cd60735fd9cd46e6