reimplement IOCP notification logic in C++
Summary:
Follow up of D46152196, which made possible to bridge coroutines between folly and python on Windows.

This diff is addressing two issues:
1. semifutures crashing on acquiring a mutex
2. C++ trying to invoke cython code after python is shutting down

I've addressed (2) by rewriting the Windows API code from cython to C++, which also removes the awkward imports of cython headers in C++.

To solve (1), I've removed the mutex and handled all the logic with `folly::ConcurrentHashMap`.

This fixes semifutures and, in part, driving on tear down, which doesn't crash anymore but somehow it hangs.

Fibers are still broken on Windows. `test_fibers` crashes with this stack trace: P776440251.

Reviewed By: yfeldblum

Differential Revision: D46971991

fbshipit-source-id: d96ded9dda598cc198096617070432edaf657759